refactor: split database persistence repositories
