+++ backend/repositories/database/__init__.py
@@ -1,1 +1,13 @@
-"""Blog and application database repositories."""
+"""SQLite runtime and application persistence repositories."""
+
+from .books import BookRepository
+from .documents import DocumentRepository
+from .history import HistoryRepository
+from .runtime import SQLiteRuntime
+
+__all__ = [
+    "BookRepository",
+    "DocumentRepository",
+    "HistoryRepository",
+    "SQLiteRuntime",
+]
